Updated validator client StateParams definition
